manufacturer code cleanup

+++ src/main/java/com/nulp/mobilepartsshop/core/service/manufacturer/ManufacturerServiceImpl.java
@@ -2,15 +2,14 @@ package com.nulp.mobilepartsshop.core.service.manufacturer;
 
 import com.nulp.mobilepartsshop.api.v1.manufacturer.service.ManufacturerLogoService;
 import com.nulp.mobilepartsshop.api.v1.manufacturer.service.ManufacturerService;
-import com.nulp.mobilepartsshop.core.enums.ImageType;
 import com.nulp.mobilepartsshop.core.entity.manufacturer.Manufacturer;
 import com.nulp.mobilepartsshop.core.entity.manufacturer.ManufacturerLogo;
+import com.nulp.mobilepartsshop.core.enums.ImageType;
 import com.nulp.mobilepartsshop.core.repository.manufacturer.ManufacturerRepository;
 import com.nulp.mobilepartsshop.exception.image.ImageDeleteException;
 import com.nulp.mobilepartsshop.exception.image.ImageGetInputStreamException;
 import com.nulp.mobilepartsshop.exception.image.ImageSaveException;
 import com.nulp.mobilepartsshop.exception.image.ImageStoreException;
-import com.nulp.mobilepartsshop.exception.manufacturer.ManufacturerNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
@@ -43,32 +42,45 @@ public class ManufacturerServiceImpl implements ManufacturerService {
     }
 
     @Override
-    public Manufacturer createManufacturer(String name, MultipartFile logo, ImageType imageType) throws ImageSaveException {
+    public Manufacturer createManufacturer(
+            String name,
+            MultipartFile logo,
+            ImageType imageType
+    ) throws ImageSaveException {
         Manufacturer manufacturer = new Manufacturer();
         manufacturer.setName(name);
-        final ManufacturerLogo manufacturerLogo = manufacturerLogoService.createManufacturerLogo(manufacturer, logo, imageType);
+        ManufacturerLogo manufacturerLogo = manufacturerLogoService.createManufacturerLogo(manufacturer, logo, imageType);
         manufacturer.setLogo(manufacturerLogo);
         return manufacturerRepository.save(manufacturer);
     }
 
     @Override
-    public Manufacturer updateManufacturer(Long id, String name, MultipartFile newLogoImage, ImageType imageType)
-            throws ManufacturerNotFoundException, ImageStoreException {
-        Manufacturer manufacturer = manufacturerRepository.findById(id)
-                .orElseThrow(() -> new ManufacturerNotFoundException("Manufacturer not found with id: " + id));
+    public Optional<Manufacturer> updateManufacturer(
+            Long id,
+            String name,
+            MultipartFile newLogoImage,
+            ImageType imageType
+    ) throws ImageStoreException {
+        Optional<Manufacturer> optionalManufacturer = manufacturerRepository.findById(id);
+        if (optionalManufacturer.isEmpty()) {
+            return optionalManufacturer;
+        }
+        Manufacturer manufacturer = optionalManufacturer.get();
         manufacturer.setName(name);
         if (!newLogoImage.isEmpty()) {
             manufacturerLogoService.updateManufacturerLogo(manufacturer.getLogo(), newLogoImage, imageType);
         }
-        return manufacturerRepository.save(manufacturer);
+        return Optional.of(manufacturerRepository.save(manufacturer));
     }
 
     @Override
-    public void deleteManufacturer(Long id) throws ManufacturerNotFoundException, ImageDeleteException {
-        Manufacturer manufacturer = manufacturerRepository.findById(id).orElseThrow(
-                () -> new ManufacturerNotFoundException("Manufacturer not found with id: " + id)
-        );
-        manufacturerLogoService.deleteManufacturerLogo(manufacturer.getLogo());
+    public boolean deleteManufacturer(Long id) throws ImageDeleteException {
+        Optional<Manufacturer> optionalManufacturer = manufacturerRepository.findById(id);
+        if (optionalManufacturer.isEmpty()) {
+            return false;
+        }
+        manufacturerLogoService.deleteManufacturerLogo(optionalManufacturer.get().getLogo());
         manufacturerRepository.deleteById(id);
+        return true;
     }
 }
